+++ src/dev-tools.ts
@@ -22,7 +22,7 @@ export function enableDevTools() {
 
     const devTools = devToolsEnabled ? (window as any).__REDUX_DEVTOOLS_EXTENSION__.connect() : undefined;
 
-    const mutators: (() => void)[] = [];
+    // const mutators: (() => void)[] = [];
 
     devTools.init(getGlobalTyduxState());
 
@@ -45,8 +45,8 @@ export function enableDevTools() {
 
     globalStateChanges$
             .subscribe(event => {
-                const mutator = !_.isNil(event.boundMutator) ? event.boundMutator : _.noop;
-                mutators.push(mutator);
+                // const mutator = !_.isNil(event.boundMutator) ? event.boundMutator : _.noop;
+                // mutators.push(mutator);
                 const meta = event.duration !== undefined ? ` (${event.duration}ms)` : "";
                 const action = {
                     ...event.action,